apps/api-server/pyproject.toml
[project]
name = "skill-hub-api"
version = "0.1.0"
description = "FastAPI backend for Skill Hub"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "alembic>=1.15.2,<2.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "boto3>=1.37.28,<2.0.0",
  "fastapi>=0.115.12,<1.0.0",
  "gunicorn>=23.0.0,<24.0.0",
  "markdown-it-py>=3.0.0,<4.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "psycopg[binary]>=3.2.6,<4.0.0",
  "python-jose>=3.4.0,<4.0.0",
  "python-multipart>=0.0.20,<0.1.0",
  "pyyaml>=6.0.2,<7.0.0",
  "redis>=5.2.1,<6.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0",
  "ruff>=0.11.2,<1.0.0"
]

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py314"
